// front/src/types/product.ts
interface product {
  CATEGORY: null;
  CHG_DATE: string;
  CHG_ID: string;
  DELETE_YN: string;
  NOTICE: null;
  PRICE: string;
  PROD_ID: number;
  RATE: string;
  REG_DATE: string;
  REG_ID: string;
  SALE_YN: string;
  SUB_TITLE: string;
  S_PRICE: null;
  S_RATE: null;
  TITLE: string;
  TUMB_IMG: string;
}

export default product;